dragonfly: Replaced all float32s with float64s internally. Even though Minecraft doesn't go over a precision of a float32, having a float64 server-side will allow for more accurate calculations. Additionally, Go's default typed float is a float64, making it easier to use those over float32s, for example with math functions that only operate on float64s.

+++ dragonfly/item/pickaxe.go
@@ -35,7 +35,7 @@ func (p Pickaxe) MaxCount() int {
 }
 
 // AttackDamage returns the attack damage of the pickaxe.
-func (p Pickaxe) AttackDamage() float32 {
+func (p Pickaxe) AttackDamage() float64 {
 	return p.Tier.BaseAttackDamage + 1
 }
